requirements.txt
pytest==8.4.2
allure-pytest==2.15.0
selene==2.0.0rc9
allure-python-commons==2.15.0
config==0.5.1
requests==2.32.5
Appium-Python-Client==5.2.4
selenium==4.36.0
python-dotenv==1.1.1